modularize knock adc for h7 only:uaefi
